requirements.txt
python-telegram-bot==4.3.2
lxml==3.6.0
requests==2.10.0
boto==2.40.0